README update and commenting

+++ munch.py
@@ -9,29 +9,14 @@ class Installer:
         self.dst = dst
         self.errors = []
         self.make_dir()
-        #self.setup_terminal_notifier()
         self.setup_munch()
 
     def make_dir(self):
         if not os.path.exists(self.dst):
             os.makedirs(self.dst)
 
-    def setup_terminal_notifier(self):
-        """Install Terminal Notifier with Homebrew"""
-        print "Installing Terminal Notifier..."
-        print "-------------------------------------"
-        try:
-            subprocess.call(['brew', 'install', 'terminal-notifier'])
-            print "-------------------------------------"
-            return True
-        except:
-            print "-------------------------------------"
-            print "There was a problem installing Terminal Notifier. Please try again or install manually."
-            self.problem.append('Problem installing Terminal Notifier.')
-            return False
-
     def setup_munch(self):
-        """Install Munch to ~/.bin/"""
+        """Install Munch to /usr/local/bin/"""
         print "Installing Munch..."
         try:
             subprocess.call(['chmod', '+x', self.src])
@@ -39,18 +24,20 @@ class Installer:
             print 'Munch installed successfully.'
             return True
         except:
-            self.problem.append('Problem installing Munch.')
+            self.errors.append('There was a problem installing Munch.')
             return False
 
 
 class Munch:
-    def __init__(self, branch = 'master', name = False):
+    def __init__(self, branch, name):
         self.branch = branch
-        self.url = 'https://github.com/bigspring/lunchbox/archive/{0}.zip'.format(branch)
+        self.url = 'https://github.com/bigspring/lunchbox/archive/{}.zip'.format(branch)
         self.lunchbox_zip = 'lunchbox.zip'
         self.composer_filename = 'composer.json'
         self.composer_data = False
         self.project_name = name
+        self.files_to_delete = ['composer.lock', self.lunchbox_zip, self.composer_filename]
+        self.dirs_to_delete = ['vendor']
         self.execute()
 
     def execute(self):
@@ -66,21 +53,32 @@ class Munch:
         self.munch_lunch()
         print "-------------------------------------"
         print "Lunch eaten."
+        print "Cleaning up mess..."
         self.cleanup()
+        print "Mess cleaned up."
 
     def fetch_lunchbox(self):
         """Download the Lunchbox repo as a zip file"""
-        urllib.urlretrieve(self.url, self.lunchbox_zip)
+        try:
+            urllib.urlretrieve(self.url, self.lunchbox_zip)
+        except:
+            print "There was a problem getting the repo."
+            print "Please ensure that the URI '{}' exists and that the".format(self.url)
+            print "monolith branch specified in composer.json also exists."
 
     def open_lunchbox(self):
         """Prepare Lunchbox files"""
         zf = zipfile.ZipFile(self.lunchbox_zip, 'r')
-        self.composer_data = zf.read('lunchbox-{0}/'.format(self.branch) + self.composer_filename)
-        if self.project_name == False:
-            self.project_name = raw_input('Enter the project name: ')
+        self.composer_data = zf.read('lunchbox-{}/'.format(self.branch) + self.composer_filename)
+        name_prompt = 'Enter the project name: '
+        if not self.project_name:
+            self.project_name = raw_input(name_prompt)
             while os.path.isdir(os.getcwd() + '/' + self.project_name):
                 print "A directory with that name already exists."
-                project_name = raw_input('Enter the project name: ')
+                self.project_name = raw_input(name_prompt)
+        else:
+            self.project_name = os.path.dirname(os.getcwd())
+        print "Project name is " + self.project_name
         self.composer_data = self.composer_data.replace('LUNCHBOX_DIR', self.project_name)
         file = open(self.composer_filename, 'w')
         file.write(self.composer_data)
@@ -88,20 +86,28 @@ class Munch:
 
     def munch_lunch(self):
         """Run composer install"""
-        subprocess.call(['composer', 'install'])
+        try:
+            subprocess.call(['composer', 'install'])
+        except:
+            print "There was a problem. Please ensure that composer is installed and functioning."
+            print "Exiting..."
+            sys.exit(2)
 
     def cleanup(self):
-        """Cleanup unrequired files"""
-        for file in [self.composer_filename, 'composer.lock', self.lunchbox_zip]:
-            os.remove(os.getcwd() + '/' + file)
-        shutil.rmtree('vendor')
+        """Cleanup unrequired files/directories"""
+        for file in self.files_to_delete:
+            if os.path.exists(os.getcwd() + '/' + file):
+                os.remove(os.getcwd() + '/' + file)
+        for dir in self.dirs_to_delete:
+            if os.path.exists(dir):
+                shutil.rmtree(dir)
 
 def correct_usage():
     """Print usage information"""
     print "usage: munch [-h], [-n projectname] [-b]"
     print "\t-h --help\tPrint correct usage and options"
     print "\t-n --name\tProvide project name"
-    print "\t-b --branch\tBranch to pull from"
+    print "\t-b --branch\tBranch from which to pull lunchbox"
 
 def install(src, dst):
     """Run installer"""
@@ -113,15 +119,17 @@ def install(src, dst):
 
 if __name__ == '__main__':
 
+    sys.exit(2)
+
+    # define defaults
     bin_path = '/usr/local/bin/munch'
     src = os.path.realpath(__file__) # current filepath
     dst = '/usr/local/bin/' # bin directory
     branch = 'master'
     project_name = False
-    installed = False
 
     try:
-        opts, args = getopt.getopt(sys.argv[1:], 'hn:b:', ['help', 'name=', 'branch='])
+        opts, args = getopt.getopt(sys.argv[1:], 'hn:b:u:', ['help', 'name=', 'branch='])
     except getopt.GetoptError:
         correct_usage()
         sys.exit(2)
